requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
pymupdf==1.26.7
watchdog==6.0.0
psutil==7.1.3
python-docx==1.1.0
openpyxl==3.1.2
requests==2.31.0
PySide6
SpeechRecognition
pyttsx3
reportlab
openai-whisper
mmh3
